CODEBASE: Refactor ImportSave component
