--- src/ui/React/ImportSave/index.ts
@@ -1,1 +1,0 @@
-export * from "./ImportSave";
